Wed, Mar 9, 2022, 7:32 PM +00:00

+++ finance/app.py
@@ -80,13 +80,13 @@ def buy():
 
         result = lookup(request.form.get("symbol"))
 
-        price = result[price]
+        price = int(result[price])
         name = result[name]
         symbol = result[symbol]
         shares = int(request.form.get("shares"))
         user_id = session["user_id"]
-        cash = int(db.execute("SELECT cash FROM users WHERE id = ?", user_id)[0]['cash']
-        new_cash = int(cash - price * shares)
+        cash = int(db.execute("SELECT cash FROM users WHERE id = ?", user_id)[0]['cash'])
+        new_cash = cash - price * shares
 
         if new_cash < 0:
             return apology("You gonna need more cash for that")
